frontend: extract useRepoPicker hook from RepoPickerStep
Splits the fetch/select logic out from submit behavior so Library's
AddReposPanel can reuse it with a different submit flow.
